requirements.txt
streamlit==1.32.2
paho-mqtt>=2.1.0
pandas>=2.0.0
Pillow>=10.0.0
streamlit-autorefresh==1.0.1